requirements.txt
-e .
aws_cdk.aws_lambda
aws_cdk.aws_dynamodb
aws_cdk.aws_apigateway
